requirements.txt
Pillow
Gymnasium==0.26.3
networkx==3.1
numpy==1.24.3
pandas==2.0.3
pygame==2.4.0
ray[rllib,default]==2.6.3
torch>=1.13.1
torch_geometric>=2.5.3
tqdm>=4.64.1
